+++ lib/code_on_the_rocks.dart
@@ -3,21 +3,24 @@ library code_on_the_rocks;
 import 'package:flutter/material.dart';
 
 /// A builder signature that returns a Widget and provides its children with a ViewModel
-typedef ModelBuilder<TViewModel> = Widget Function(
-    BuildContext context, TViewModel model);
+typedef _ModelBuilder<TViewModel> = Widget Function(BuildContext context, TViewModel model);
 
 /// A ViewModelBuilder is a StatefulWidget that builds a ViewModel of type T
 /// This is the widget you will add to your widget tree
 abstract class ViewModelBuilder<TViewModel> extends StatefulWidget {
   const ViewModelBuilder({Key? key, required this.builder}) : super(key: key);
-  final ModelBuilder<TViewModel> builder;
+  final _ModelBuilder<TViewModel> builder;
 }
 
 /// Get the nearest ViewModel of type T
 /// This is a convenience method that can be used to get the ViewModel from a BuildContext
-T getModel<T>(BuildContext context) => (context
-    .dependOnInheritedWidgetOfExactType<ViewModelProvider<ViewModel<T>>>()!
-    .state) as T;
+T getModel<T>(BuildContext context) {
+  try {
+    return (context.dependOnInheritedWidgetOfExactType<ViewModelProvider<ViewModel<T>>>()!.state) as T;
+  } catch (e) {
+    throw Exception("Could not find ViewModel of type $T above this widget. Make sure you have added a matching ViewModelBuilder to the widget tree.");
+  }
+}
 
 /// A ViewModel is a State object that can be used to store data and update the UI
 /// When the ViewModel calls setState, the UI will be updated
@@ -55,7 +58,7 @@ abstract class ViewModel<T> extends State<ViewModelBuilder<T>> {
 /// A ModelWidget is a StatelessWidget that provides its children with a ViewModel
 class ModelWidget<TViewModel extends ViewModel> extends StatelessWidget {
   const ModelWidget({Key? key, required this.builder}) : super(key: key);
-  final ModelBuilder<TViewModel> builder;
+  final _ModelBuilder<TViewModel> builder;
 
   @override
   Widget build(BuildContext context) {
@@ -70,9 +73,7 @@ class ModelWidget<TViewModel extends ViewModel> extends StatelessWidget {
 /// A ViewModelProvider is an InheritedWidget that provides its children with a ViewModel
 /// It is returned by the ViewModel's build method and should not be edited
 class ViewModelProvider<TViewModel extends ViewModel> extends InheritedWidget {
-  const ViewModelProvider(
-      {Key? key, required Widget child, required this.state})
-      : super(key: key, child: child);
+  const ViewModelProvider({Key? key, required Widget child, required this.state}) : super(key: key, child: child);
   final TViewModel state;
 
   @override