+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.1
+- [Changed] Made ModelBuilder TypeDef private
+- [Added] Added useful error message when inherited ViewModel cannot be found
+- **Date**: 8-25-2023
+
 ## 0.1.0
 - [Added] Added ModelWidget class
 - [Changed] Updated readme to show ModelWidget usage